fix(tracing): Remove `isInstanceOf` check in Hub constructor (#5046)
fix a bug which assigned the wrong Hub instance to a Transaction in the Transaction constructor. Remove the isInstanceOf check which incorrectly failed when passing a user-created Hub the constructor. Overall, simplify the hub initialization by deciding in the constructor if the passed Hub should be used. As a fallback, the Hub returned from getCurrentHub is used.

Co-authored by: @lobsterkatie