ref(node): Explicitly pass down node transport options
We enforce the transport types by assigning the generic when typing the
Node Client class (we use `ClientOptions<NodeTransportOptions>`).

To further refine the transport types, we remove the transport options
from the top level of the sdk init and instead require users to pass it
in under the `transportOptions` key explicitly.