ref(build): Use rollup/sucrase for non-watch npm builds (#5035)
This switches our `build` and `build:dev` yarn scripts to use the new rollup/sucrase build process.

This is the culmination of a number of previous changes, whose highlights include:
- https://github.com/getsentry/sentry-javascript/pull/4724, which made building types files a separate build process
- https://github.com/getsentry/sentry-javascript/pull/4895, which updated the SDK to use TS 3.8.3
- https://github.com/getsentry/sentry-javascript/pull/4926, which removed use of the `Severity` enum
- https://github.com/getsentry/sentry-javascript/pull/5005, which switch our default tsconfig to target es6
- https://github.com/getsentry/sentry-javascript/pull/4992, which added the Sucrase plugin, some helper functions, and the `yarn build:rollup` script
- https://github.com/getsentry/sentry-javascript/pull/4993, which added rollup plugins to use `var` rather than `const` and clean up the built code in various ways
- https://github.com/getsentry/sentry-javascript/pull/5022, which applied the same `const`-to-`var` translation to tests
- https://github.com/getsentry/sentry-javascript/pull/5023, which added the ability to change injected polyfills into imports

The result is that, as of this PR, we will no longer use `tsc` to transpile or down-complile our code when building npm packages. Instead, we will be using Rollup to handle making our code CJS-friendlly and Sucrase to handle the transpilation from TS to JS. The main advantages of this change are:
- It forced us to do a lot of updating, centralizing, and cleanup of our tooling, not just for build but also for testing and linting.
- It brings our CDN build process and our npm build process much more in line with each other, for easier maintainability.
- It gives us more control over the eventual output, because we have access to a whole internet full of Rollup plugins (not to mention the ability to make our own), rather than being constrained to tsconfig options. (Plugins also allow us to interact with the code directly.)
- It speeds up our builds fairly significantly. I ran a number of trials in GHA of running `yarn build:dev` at the top level of the repo. Before this change, the average time was ~150 seconds. After this change, it's about half that, roughly 75 seconds.

Because of the switch in tooling, the code we publish is going to be slightly different. In order to make sure that those differences weren't going to be breaking, I built each package under the old system and under the new system, ran a `git diff`, and checked every file, both CJS and ESM, in every package affected by this change. The differences (none of which affect behavior or eventual bundle size by more than a few bytes in each direction), fell into a few categories:

- Purely cosmetic changes, things like which comments are retained, the order of imports, where in the file exports live, etc.
- Changes to class constructors, things like not explicitly assigning `undefined` to undefined attributes, using regular assignment rather than `Object.defineProperty` for attributes which are assigned values, and splitting some of those assignments off into helper functions. 
- Changes related to the upgrade to ES6 and dropping of support for Node 6, things like not polyfilling object spread or async/await

While this represents the most significant part of the overall change, a few outstanding tasks remain:

- Making this same switch in `build:watch`
- Parallelizing the builds, both locally and in CI
- Perhaps applying the new process to our CDN bundle builds
- Generalized cleanup

These will all be included in separate PRs, some in the immediate future and some in the hopefully-not-too-distant short-to-medium term.