chore(dev): Remove redundant `prepublishOnly` build step in CI (#5083)
Based both on the description in the comment above it and the restrictions on when it is supposed to run, it appears that the `prepublishOnly` step of the `Build` job is meant to do the same thing as the `pack` step of the `Upload Artifacts` job. "Meant" is the operative word, however, because in reality, it doesn't actually do anything - though it calls the repo-level `prepublishOnly` script, which calls the `prepublishOnly` script in every package that has one... no package has one, so it's a no-op.

This removes the step and the associated repo-level yarn script, since they seem to have been replaced by the `build:npm` script called in `Upload Artifacts`.