feat(dev): Add default Node version for CI jobs (#5081)
In our CI workflow, we currently use version 1 of GHA's `setup-node` action, whose default Node version is 10. Though we could (and should) update to version 3 of the action, whose default is Node 16, the `setup-node` docs[1] recommend explicitly setting a version for greater control and clarity. This does so, by adding the default Node version as an env variable and using it in all jobs which don't otherwise specify their version.

[1] https://github.com/actions/setup-node#usage