fix(types): Make `@sentry/types` a runtime dependency in gatsby and nextjs (#5086)
In order that it be installed with the SDK, we make the types package a runtime/regular dependency of all of our other packages, so it's there for anyone using the SDK in a TS app. In the gatsby and nextjs packages, though, it's listed as a dev dependency, and therefore won't be installed. This corrects that by moving it to be a regular dependency in both packages.