fix(dev): Restrict size check action to PRs (#5082)
The size check action we use in our main CI workflow will throw an error if run outside of the context of a PR (as when running CI manually, for example). As a result, the overall run is always marked as a failure, even when all other jobs passed. This restricts the action to running on PRs, to avoid that problem.