chore(build): Move aws lambda build step (#5123)
This moves, without making any changes, the aws lambda build job to live with the other build job in our CI workflow. (I'll admit that it's not totally obvious, but there is in fact a vague organizing principle to that doc.) Extracted from another PR to make it easier to see, in that PR, what actual changes are being made.