feat(build): Allow for multiple bundle entrypoints (#5143)
This adds the ability to create more than one bundle with a single bundle rollup config, by specifying multiple entrypoints. Because multiple files can now be output, the way they're named also needed to be switched, from providing a static string to providing a function to determine the name of each bundle. Finally, the name of the option (`input`) was changed to match the corresponding option used in the building of npm packages (`entryponts`).