feat(build): Add option to build only specified bundle variants (#5145)
There are situations in which we really only need one version of a bundle rather than all versions, but right now there's no way to build one without building all of them. This adds the option to specify which bundles you need, defaulting to the current behavior of building all three types (regular, minified, and minified with logging) if particular ones aren't requested.