feat(tracing): Add empty baggage header propagation to outgoing requests (#5133)
Adds the propagation of an "empty" baggage header. The word "empty" is however kind of misleading as the header is not necessarily empty. In order to comply with the baggage spec, as of this patch, we propagate incoming (3rd party) baggage to outgoing requests. The important part is that we actually add the `baggage` HTTP header to outgoing requests which is a breaking change in terms of CORS rules having to be adjusted.

We don't yet add `sentry-` baggage entries to the propagated baggage. This will come in a follow up PR which does not necessarily have to be part of the initial v7 release as it is no longer a breaking change. 

Overall, this is heavily inspired from #3945 (thanks @lobsterkatie for doing the hard work)

More specifically, this PR does the following things:

1. Extract incoming baggage headers and store them in the created transaction's metadata. Incoming baggage data is intercepted at:
* Node SDK: TracingHandler
* Serverless SDK: AWS wrapHandler
* Serverless SDK: GCP wrapHttpFunction
* Next.js: SDK makeWrappedReqHandler
* Next.js: SDK withSentry
* BrowserTracing Integration: by parsing the `<meta>` tags (analogously to the `sentry-trace` header)

2. Add the extracted baggage data to outgoing requests we instrument at:
* Node SDK: HTTP integration
* Tracing: instrumented Fetch and XHR callbacks

Co-authored-by: Luca Forstner <luca.forstner@sentry.io>