fix(dev): Add placeholder tsconfigs for tests (#5169)
It is well-documented (see 1-6 below) that VSCode doesn't support setups like ours, where multiple tsconfig files coexist in a single directory. Strangely, though, it is only recently that this has become a problem, with VSCode at random intervals forgetting that it's ever heard of `expect` or `describe` (because it's not seeing `tsconfig.test.json`, but taking a while to realize it).

There is an open issue[7] tracking the addition of support for this, but it's been open for a long time, with little movement. In the meantime, this solves the problem by adding placeholder `test/tsconfig.json` files to each package, each pointing to its corresponding `tsconfig.test.ts` file. I went with this approach over simply moving and renaming the existing test tsconfigs because this allows us to stay consistent in having all flavors of tsconfig for a package live at the package root level, and provides an easy way to reverse this workaround, should VSCode ever fix the underlying problem.

[1] https://github.com/angular/angular-cli/issues/5175
[2] https://github.com/microsoft/TypeScript/issues/49210
[3] https://github.com/microsoft/vscode/issues/107750
[4] https://github.com/Microsoft/vscode/issues/12463
[5] https://github.com/mseag/bible-karaoke/issues/175
[6] https://github.com/microsoft/TypeScript/issues/8435
[7] https://github.com/microsoft/TypeScript/issues/33094