feat(tracing): Propagate environment and release values in baggage Http headers (#5193)
Add the sentry-environment and sentry-release values to the baggage HTTP headers of outgoing requests. We add these values to baggage as late as possible with the rationale that other baggage values (to be added in the future) will not be available right away, e.g. when intercepting baggage from incoming requests.

As a result, the flow described below happens, when the first call to span.getBaggage is made (e.g. in callbacks of instrumented outgoing requests):

1. In span.getBaggage, check if there is baggage present in the span (in which case it was intercepted from incoming headers/meta tags) and it is empty (or only includes 3rd party data) OR if there is no baggage yet
2. In both of these cases, populate the baggage with Sentry data (and leave 3rd party content untouched). Else do nothing
3. Add this baggage to outgoing headers (and merge with possible 3rd party header)

Additionally, add and improve tests to check correct handling and propagation of baggage