chore(remix): Update package to 7.x structure (#5266)
This is a follow-up to https://github.com/getsentry/sentry-javascript/pull/5256 and https://github.com/getsentry/sentry-javascript/pull/5264, which added a skeleton remix SDK to the repo. The template that was used to create the new package had parts of the 6.x package structure and configuration still in it. This fixes that by bringing everything up to date to match our other packages.